requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
deep-translator>=1.11.4
ollama>=0.3.0

